Configure testing branch.
- Create jest.config and jest.setup.
- Fix folder structure to follow bulletproof.
- Remove routes files and directly import routes in app.jsx.
- Add react-router and redux configuration for testing environment.
