+++ src/Pages/Registration/Components/Form/RegistrationForm.jsx
@@ -3,12 +3,11 @@ import { useForm } from "react-hook-form";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 
-import { loginSuccess } from "../../../Stores/Actions/auth";
+import { loginSuccess } from "../../../../Stores/Actions/auth";
+import { LoginFormValidationSchema } from "../../../Login/Validation/LoginValidation";
+import { createNewUser } from "../../Api/RegistrationMethods";
 
-import { loginUser } from "../Api/LoginMethods";
-import { LoginFormValidationSchema } from "../Validation/LoginValidation";
-
-const LoginForm = () => {
+const RegistrationForm = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
@@ -27,7 +26,7 @@ const LoginForm = () => {
   });
 
   const onSubmit = async (user) => {
-    const response = await loginUser(user);
+    const response = await createNewUser(user);
 
     if (response) {
       const newUser = {
@@ -40,7 +39,7 @@ const LoginForm = () => {
 
       dispatch(loginSuccess(newUser));
 
-      alert("Logged in Successfully");
+      alert("Signed Up Successfully");
 
       navigate(`/${newUser.id}`);
     }
@@ -53,7 +52,7 @@ const LoginForm = () => {
       onSubmit={handleSubmit(onSubmit)}
       className="flex flex-col gap-4 items-center justify-center w-[500px] px-10 py-12 border border-slate-200"
     >
-      <h1 className="font-medium text-lg mb-5">LOGIN</h1>
+      <h1 className="font-medium text-lg mb-5">REGISTER</h1>
 
       <div className="form-control w-full max-w-xs">
         <label className="label">
@@ -101,4 +100,4 @@ const LoginForm = () => {
   );
 };
 
-export default LoginForm;
+export default RegistrationForm;
