--- src/Shared/Utils/ProtectedRoute.jsx
@@ -1,13 +1,0 @@
-import { Navigate, useLocation } from "react-router-dom";
-
-const ProtectedRoute = ({ children }) => {
-  const token = localStorage.getItem("token");
-  let location = useLocation();
-
-  if (!token) {
-    return <Navigate to="/" state={{ from: location }} replace />;
-  }
-  return children;
-};
-
-export default ProtectedRoute;
